text-summarization-pipeline/requirements.txt
transformers==4.40.0
datasets==2.19.0
torch>=2.2.0
rouge-score==0.1.2
numpy>=1.26.0
tqdm>=4.66.0
pyyaml>=6.0.1
matplotlib>=3.8.0
py7zr